+++ src/php/Compiler/Domain/Analyzer/TypeAnalyzer/PhpKeywords.php
@@ -75,5 +75,6 @@ final class PhpKeywords
         'var',
         'while',
         'xor',
+        'yield',
     ];
 }